refactor(commands): extract /config command into config_cmd.rs
